+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+##### 06.01.2025
+
+> Feature: Approach page UI development
+> Fetaure: About us page UI development
+> Feature: Contact Us page UI and contact form integrations
+
 ##### 05.01.2025
 
 > Feature: Services page UI additions and updates